Restructured folders/pages for next js Routing renamed dashboards to index's for routing implemented withAuth on dash boards SignIn now redirects users to correct dashboard depending on their role Implemented functions to fetch user data/roles when needed Updated links due to new file structures
