+++ frontend/components/hooks/CurrentUserContext.js
@@ -1,0 +1,42 @@
+import React from "react";
+import axios from "axios";
+
+export const CurrentUserContext = React.createContext();
+
+export const CurrentUserProvider = ({ children }) => {
+  const [currentUser, setCurrentUser] = React.useState(null);
+
+  const instance = axios.create({
+    withCredentials: true,
+  });
+
+  const fetchCurrentUser = async () => {
+    try {
+      let response = await instance.get("http://localhost:8080/api/customers");
+
+      setCurrentUser(response);
+    } catch (error) {
+      console.error(error);
+    }
+  };
+
+  return (
+    <CurrentUserContext.Provider value={{ currentUser, fetchCurrentUser }}>
+      {children}
+    </CurrentUserContext.Provider>
+  );
+};
+
+export const useCurrentUser = () => React.useContext(CurrentUserContext);
+
+// can be used in other components to retrieve user info
+// const { currentUser, fetchCurrentUser } = useCurrentUser();
+// React.useEffect(() => {
+//   const getCurrentUser = async () => {
+//     fetchCurrentUser();
+//   };
+
+//   getCurrentUser();
+// }, []);
+
+// console.log(currentUser);
