+++ frontend/components/router/withAuth.js
@@ -1,21 +1,25 @@
 import { useRouter } from "next/router";
-import { useEffect } from "react";
+import { useEffect, useState } from "react";
+import { checkRole, checkLogin } from "../hooks/checkUser";
 
 // could possibly add auth to determine if there is a token in the system as well.
 const withAuth = (WrappedComponent, allowedRoles) => {
   const Authenticate = (props) => {
     const router = useRouter();
+    const [isAuthorized, setIsAuthorized] = useState(false);
 
     useEffect(() => {
       // Check if the user has the necessary role to access the page
-      if (!allowedRoles.includes("Admin")) {
+      if (allowedRoles.includes(checkRole())) {
         // replace "Admin" with user.role
+        setIsAuthorized(true);
+      } else {
         router.push("/unauthorized");
       }
-    }, ["Admin"]); // replace "Admin" with user.role
+    }, []); // replace "Admin" with user.role
 
     // Render the wrapped component if the user has the necessary role
-    return allowedRoles.includes("Admin") ? ( // replace "Admin" with user.role
+    return isAuthorized ? ( // replace "Admin" with user.role
       <WrappedComponent {...props} />
     ) : null;
   };
