+++ frontend/package.json
@@ -21,6 +21,7 @@
     "dayjs": "^1.11.7",
     "install": "^0.13.0",
     "js-cookie": "^3.0.1",
+    "jwt-decode": "^3.1.2",
     "mui-nested-menu": "^3.2.0",
     "next": "^13.3.0",
     "npm": "^9.6.4",
