+++ frontend/pages/_app.js
@@ -9,6 +9,7 @@ import createEmotionCache from "../src/createEmotionCache";
 import Developer from ".";
 import { useRouter } from "next/router";
 import { Switch, Route } from "react-router-dom";
+import { CurrentUserProvider } from "../components/hooks/CurrentUserContext";
 
 // Client-side cache, shared for the whole session of the user in the browser.
 const clientSideEmotionCache = createEmotionCache();
@@ -25,6 +26,9 @@ export default function MyApp(props) {
         {/* CssBaseline kickstart an elegant, consistent, and simple baseline to build upon. */}
         <CssBaseline />
         <Component {...pageProps} />
+        {/* <CurrentUserProvider>
+          ... component
+        </CurrentUserProvider> */}
       </ThemeProvider>
     </CacheProvider>
   );
